+++ src/test/suite/extension.test.ts
@@ -5,18 +5,18 @@ suite('Extension Test Suite', () => {
     vscode.window.showInformationMessage('Start all tests.');
 
     test('Extension should be present', () => {
-        assert.ok(vscode.extensions.getExtension('YourPublisherName.literate'));
+        assert.ok(vscode.extensions.getExtension('kwesinavilot.changescribe'));
     });
 
     test('Should activate extension', async () => {
-        const ext = vscode.extensions.getExtension('YourPublisherName.literate');
+        const ext = vscode.extensions.getExtension('kwesinavilot.changescribe');
         await ext?.activate();
         assert.ok(ext?.isActive);
     });
 
     test('Should register generate changelog command', () => {
         return vscode.commands.getCommands(true).then((commands) => {
-            assert.ok(commands.includes('literate.generateChangelog'));
+            assert.ok(commands.includes('changeScribe.generateChangelog'));
         });
     });
 });